requirements.txt
appdirs==1.4.3
config==0.3.9
diff-match-patch==20121119
Django==1.11.7
django-admin-view-permission==0.6
django-chosen==0.1
django-filter==1.0.1
django-import-export==0.5.1
django-nested-inline==0.3.7
django-simple-history==1.8.2
djangorestframework==3.6.2
drf-extensions==0.3.1
ldap3==2.2.2
Markdown==2.6.8
packaging==16.8
pyasn1==0.2.3
pyldap==2.4.25.1
pyparsing==2.2.0
python-memcached==1.58
pytz==2016.10
pyvmomi==6.5
requests==2.13.0
six==1.10.0
tablib==0.11.4
tzlocal==1.3
selenium==3.3.1
wheel==0.24.0
Sphinx==1.5.3
sphinx_bootstrap_theme==0.4.14
python-dateutil==2.6.0
django-mass-edit==3.0
beautifulsoup4==4.6.0
django-tagging==0.4.5
django-tagging-autocomplete==0.5.0
